Moving code to simplify teh pages layouts and layouts params
