+++ WelcomeCoordinator/src/main/java/com/redbooth/WelcomePagePercentLayout.java
@@ -28,6 +28,7 @@ import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.res.TypedArray;
 import android.os.Build;
+import android.support.annotation.NonNull;
 import android.support.percent.PercentRelativeLayout;
 import android.text.TextUtils;
 import android.util.AttributeSet;
@@ -53,123 +54,64 @@ public class WelcomePagePercentLayout extends PercentRelativeLayout implements W
         super(context, attrs, defStyleAttr);
     }
 
-    //@SuppressWarnings("unused")
-    //@TargetApi(Build.VERSION_CODES.LOLLIPOP)
-    //public WelcomePageViewLayout(Context context, AttributeSet attrs,
-    //                         int defStyleAttr, int defStyleRes) {
-    //    super(context, attrs, defStyleAttr, defStyleRes);
-    //}
-
     @Override
-    protected WelcomeLayoutParams generateDefaultLayoutParams() {
-        return new WelcomeLayoutParams(WelcomeLayoutParams.WRAP_CONTENT, WelcomeLayoutParams.WRAP_CONTENT);
+    protected LayoutParams generateDefaultLayoutParams() {
+        return new LayoutParams(LayoutParams.WRAP_CONTENT, LayoutParams.WRAP_CONTENT);
     }
 
     @Override
     protected boolean checkLayoutParams(ViewGroup.LayoutParams params) {
-        return params instanceof WelcomeLayoutParams;
+        return params instanceof LayoutParams;
     }
 
     @Override
-    protected WelcomeLayoutParams generateLayoutParams(ViewGroup.LayoutParams params) {
-        return new WelcomeLayoutParams(params);
+    protected LayoutParams generateLayoutParams(ViewGroup.LayoutParams params) {
+        return new LayoutParams(params);
     }
 
     @Override
-    public WelcomeLayoutParams generateLayoutParams(AttributeSet attrs) {
-        return new WelcomeLayoutParams(getContext(), attrs);
+    public LayoutParams generateLayoutParams(AttributeSet attrs) {
+        return new LayoutParams(getContext(), attrs);
     }
 
     @Override
     public List<WelcomePageBehavior> getBehaviors(WelcomeCoordinatorLayout coordinatorLayout) {
-        List<WelcomePageBehavior> result = new ArrayList<>();
-        for (int index = 0; index < getChildCount(); index++) {
-            final View view = getChildAt(index);
-            if (view.getLayoutParams() instanceof WelcomeLayoutParams) {
-                final WelcomeLayoutParams params = (WelcomeLayoutParams)view.getLayoutParams();
-                final WelcomePageBehavior behavior = params.getBehavior();
-                if (behavior != null) {
-                    behavior.setCoordinator(coordinatorLayout);
-                    behavior.setTarget(view);
-                    behavior.setPage(this);
-                    result.add(behavior);
-                }
-            }
-        }
-        return result;
+        return WelcomeBehaviorsExtract.getWelcomePageBehaviors(this, coordinatorLayout);
     }
 
-    public static class WelcomeLayoutParams extends PercentRelativeLayout.LayoutParams implements WelcomePageView.WelcomeLayoutParams {
-        public final static int NO_DESTINY_VIEW = -1;
-        private int destinyViewId = NO_DESTINY_VIEW;
-        private WelcomePageBehavior behavior;
+    public static class LayoutParams extends PercentRelativeLayout.LayoutParams implements WelcomeLayoutParams {
+        private WelcomeBehaviorLayoutParams welcomeBehaviorLayoutParams = new WelcomeBehaviorLayoutParams();
 
+        @Override
         public WelcomePageBehavior getBehavior() {
-            return behavior;
+            return welcomeBehaviorLayoutParams.behavior;
         }
 
         @Override
         public int getDestinyViewId() {
-            return destinyViewId;
+            return welcomeBehaviorLayoutParams.destinyViewId;
         }
 
-        public WelcomeLayoutParams(int width, int height) {
+        public LayoutParams(int width, int height) {
             super(width, height);
         }
 
-        public WelcomeLayoutParams(ViewGroup.LayoutParams source) {
+        public LayoutParams(ViewGroup.LayoutParams source) {
             super(source);
         }
 
-        public WelcomeLayoutParams(MarginLayoutParams source) {
+        public LayoutParams(MarginLayoutParams source) {
             super(source);
         }
 
-        public WelcomeLayoutParams(Context context, AttributeSet attrs) {
+        public LayoutParams(Context context, AttributeSet attrs) {
             super(context, attrs);
-            extractAttributes(context, attrs);
-        }
-
-        private void extractAttributes(Context context, AttributeSet attrs) {
-            final TypedArray attributes = context.obtainStyledAttributes(attrs,
-                    R.styleable.WelcomePageLayout_LayoutParams);
-            if (attributes.hasValue(R.styleable.WelcomePageLayout_LayoutParams_view_behavior)) {
-                behavior = parseBehavior(context, attributes
-                        .getString(R.styleable.WelcomePageLayout_LayoutParams_view_behavior));
-            }
-            if (attributes.hasValue(R.styleable.WelcomePageLayout_LayoutParams_destiny)) {
-                destinyViewId = attributes
-                        .getResourceId(R.styleable.WelcomePageLayout_LayoutParams_destiny, NO_DESTINY_VIEW);
-            }
-            attributes.recycle();
-        }
-
-        private WelcomePageBehavior parseBehavior(Context context, String name) {
-            WelcomePageBehavior result = null;
-            if (!TextUtils.isEmpty(name)) {
-                final String fullName;
-                if (name.startsWith(".")) {
-                    fullName = context.getPackageName() + name;
-                } else {
-                    fullName = name;
-                }
-                try {
-                    Class<WelcomePageBehavior> behaviorClazz
-                            = (Class<WelcomePageBehavior>) Class.forName(fullName);
-                    final Constructor<WelcomePageBehavior> mainConstructor
-                            = behaviorClazz.getConstructor();
-                    mainConstructor.setAccessible(true);
-                    result = mainConstructor.newInstance();
-                } catch (Exception e) {
-                    throw new RuntimeException("Could not inflate Behavior subclass " + fullName, e);
-                }
-            }
-            return result;
+            welcomeBehaviorLayoutParams.extractAttributes(context, attrs);
         }
 
         @SuppressWarnings("unused")
         @TargetApi(Build.VERSION_CODES.LOLLIPOP)
-        public WelcomeLayoutParams(PercentRelativeLayout.LayoutParams source) {
+        public LayoutParams(PercentRelativeLayout.LayoutParams source) {
             super(source);
         }
     }
